requirements.txt
Flask
pymongo
python-dotenv
gunicorn
pandas
openpyxl